requirements.txt
flask
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
numpy
tqdm
google-genai
python-jose
firecrawl